requirements-dev.txt
# Development and testing dependencies only.
# NOT shipped with the collection artifact (excluded via build_ignore in galaxy.yml).
#
# Runtime dependency: pysnmp-lextudio>=6.1,<8
# Runtime deps are documented in README.md and enforced by a fail-fast import
# guard in the module. They are NOT listed here to keep the separation clear.
#
# Install: pip install -r requirements-dev.txt
ansible-core>=2.15
ansible-lint>=6.0
antsibull-changelog>=0.26
pytest>=7.0
pytest-mock>=3.0
molecule>=6.0
molecule-plugins[docker]>=23.0
pylint>=3.0
pysnmp-lextudio>=6.1,<8
